Add print function in graph ops
